Add basic gameplay ui.
Holy magic numbers, batman!

This change adds a basic ui for playing the game in a terminal. It
accomplishes this with a new `Game` struct (in the new `game` module).
The Game struct keeps a mutable board state, the most magical cursor
ever, and a mode for modal controls. It has bindings for every feature
currently enabled by the Board model.

I swear the fact that the file's named "gamers" is unintentional.

This change rips out the TerminalPrintable trait, which I already regret
a little. I might end up adding that back in, if for no reason other
than debug. For now though, it's replaced by the functions in the
`display` module.