Fix A* search.
Turns out < and > are important to get right.

Heuristic still isn't admissable, so there's a chance this'll give a
non-optimal answer. If I notice an example of this I'll write a test
to address it. I'm curious if it's possible to come up with an
admissable but also helpful heuristic. I might have to play around
with that.